+++ src/controllers/productController.js
@@ -10,10 +10,10 @@ class ProductController {
         return await productService.getProductById(product);
     }
 
-    async addProducts(cartId, productId) {
-        return await productService.addProducts(cartId, productId);
+    async addProducts(product, user) {
+        return await productService.addProducts(product, user);
     }
-
+    
     async updateProduct(productId, productUpdate) {
         return await productService.updateProduct(productId, productUpdate);
     }