+++ src/dao/mongo/productMongoDAO.js
@@ -23,7 +23,7 @@ class ProductDao {
         return product;
     }
 
-    async addProducts(product) {
+    async addProducts(product, user) {
         const {title, description, code, price, stock, category, thumbnail} = product;
         const existingProduct = await productModel.findOne({ code });