+++ src/repositories/productRepository.js
@@ -37,17 +37,17 @@ class ProductRepository {
     }
     
 
-    async addProducts(pid) {
+    async addProducts(product, user) {
         try {
-            const product = await this.dao.addProducts(pid);        
-            return new ProductDto(product);
+          const result = await this.dao.addProducts(product, user);
+          return new ProductDto(result);
         } catch (error) {
-            throw CustomError.createError({
-              name: "Product creation error.",
-              cause: generateProductsErrorInfo(pid),
-              message: "Couldn't add Product",
-              code: EErrors.DATABASE_ERROR
-            });
+          throw CustomError.createError({
+            name: 'Product creation error.',
+            cause: generateProductsErrorInfo(product),
+            message: "Couldn't add Product",
+            code: EErrors.DATABASE_ERROR,
+          });
         }
     }